Exercise 2 - Tile set and tile map
So far, tile set and tile map are completed. Also, a helper class, tile map parser, was created to facilitate this process.
I'm not sure yet whether these tile classes are correctly implemented. The image it shows it kinda weird.
Face is behaving in the same way, which is a good sign.