Camera initial implementation
Camera exists and allows tile map scrolling.
Face objects are still static on the screen and must also be offset by the camera position. This can be made by updating the game object box center point. However, this doesn't seem to be a nice solution since we will lose the object position in the game. Another option is to allow Render() to receive an offset parameter or something like that.