Undoing abstract factory changes
I finally understood the goal of having abstract factories. It's pretty useful when several different classes can be used interchangeably and we may define them dinamically during the program execution.
For instance, we may have a "cheat interpreter" that defines dinamically which class will be used by the user: a penguin or a chicken. In order to do it, we will have two classes (`Penguin` and `Chicken`) that inherit from `GameObject`. When executing a command to change the player object, it will activate the factory, which will create the appropriate object.
So far, we do not need this complexity in the game.