Creating GameObjectManager
This class is responsible for updating and rendering game objects, as well as taking its ownership.
It is not the prettiest design ever. Actually, it may be an anti-pattern if `GameObjectManager` keeps growing.
A caveat is that objects are kept in a global container, instead of being separated in individual spaces (avoid collision, hide entire spaces, etc)