Game does not crash if trying to execute commands from deleted objects
By applying the Command pattern, `InputManager` holds only a `weak_ptr` to commands, which are created by `GameObject` instances or any other class that uses the defined command.
Then, if the owner command is deleted, `InputManager` will know that the `weak_ptr` has expired, hence, not executing the command anymore and removing it from it's control.
The whole code needs immediate refactoring though, since it was built in an experimental form.
Also, `Action` applies the Command pattern too. Would it be interesting to make them inherit from a unified interface?