Refactoring InputManager
InputManager is not a singleton anymore, since each state will have it's own input manager instance.
TileState can transition to the next state with keyboard space press.
Fixing Penguins friction with reverse movement, which was increasing instead of decreasing speed.
Removing input command registering from CommandFactory, since it is not its responsibility to register, but from the instance that is calling it.