Rendering batch images
Before this commit, each image would be independently requested to render and the GPU would attend the request immediately. This was done by invoking the `SDL_RenderPresent` method.
Now, the rendering takes place after all images requests to be rendered. A new method `RenderBatch` was created in `GraphicsAdapter` to provide this functionality.
This severely enhances the game performance. With 2500 cells, FPS is still 30 whereas previously the FPS would drop to 25 with only 250 cells.