Quadtree enhanced collision system
Quadtree is definitelly worth it. After correcting a few bugs, I was able to get about 200 particles with 30 fps.
The code is a mess and getting out of control. Nonetheless, this line of development combined with a few heuristics (limit the amount of collision detection per turn, limit the number of particles in the scene, not simulate what is out of view) is enough to make the game always run at 30 fps.