Trying and fixing "ColliderComponent not in entity X"
Apparently, this bug was happening because of combat resolution.
`entity1`, the one in `collidableEntities` vector, engaged in combat with `entity2`, the one retrieved from the quadtree. However, `entity2` was stronger than `entity1`, hence the former ate the later. In this process, `entity1` would be inserted in `deletedEntities` vector.
The problem was that the inner for loop, checking collisions for entities from collidable entities and from the quadtree, assumed collidable entities are always valid and checked only if the quadtree entity has been deleted.
Clearly, with combat system, it may happen that a collidable entity becomes invalid when it loses a combat.
Hope this was the problem indeed.