Button respond to clicks.
Now, when clicking on a button, a "callback" fires and responds to it. It would be interesting to really define this as a callback function.
Entities spawned after the button creation shows up over it. This can be solved by creating spaces in `EntityManager` and define priorities for these spaces. Then, `RenderingSystem` would iterate over the spaces to render sprites.